Refactor source routing setup
Change MTORR intervals.
